refactor: separate worker and client
