+++ zorya/worker/gcp/compute.py
